Switch to uintptr_t's to pass pointers to/from C++. (#35)
This is motivated by a similar change in mattress. It eliminates the
need to rely on pybind11 in downstream packages; as long as they can
take a uintptr_t, they can use knncolle. We add a knncolle_py.h header
with type definitions that can be used downstream for correct casting.