Switch to uintptr_t's to pass pointers to/from C++. (#35) (#1)
This is motivated by a similar change in mattress. It eliminates the need to
rely on pybind11 in downstream packages; as long as they can take a uintptr_t,
they can use knncolle. We add a knncolle_py.h header with type definitions for
correct casting of the uintptr_t; this can be obtained via include().

A side-effect of this change is that we need to implement destructors for the
pointed-to objects for each uintptr_t. This involves writing __del__ methods
for GenericIndex instances as well as creating a new Builder class. While we're
here, we  move more logic into GenericIndex to reduce repetition.